[refactor][STEP4] commander.py --> shell directory
